# filesystemDB/docs/items.swagger.yml
openapi: 3.0.0
info:
  title: FileSystem DB API
  version: 1.0.0
  description: API documentation for FileSystem DB operations
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:80/api/v1
    description: Local server

tags:
  - name: Items
    description: API to manage your items.

paths:
  /items:
    get:
      summary: Retrieve a list of items
      tags: [Items]
      responses:
        200:
          description: A list of items.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
    post:
      summary: Create a new item
      tags: [Items]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                name:
                  type: string
      responses:
        201:
          description: The created item.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string

  /items/{id}:
    put:
      summary: Update an existing item
      tags: [Items]
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: The item ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                name:
                  type: string
      responses:
        200:
          description: The updated item.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
        404:
          description: Item not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

    delete:
      summary: Delete an existing item
      tags: [Items]
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: The item ID
      responses:
        200:
          description: Item deleted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        404:
          description: Item not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string